First steps in the move towards Int indexed environments
We can index each variable by its declaration order from the point of
view of the call site, and then instead of using the scoped maps
as the environment type, we can use a better structure such as a
fingertree. This is in response to profiling which indicated that
variable resolution is a major cost centre.